Silence a number of spurious compiler warnings
    - Existing warnings are now a good proxy for areas of code smell ;-)

+++ src/audio/speaker.c
@@ -193,7 +193,7 @@ static void _speaker_update(/*bool toggled*/) {
         if (UNLIKELY(samples_buffer_idx > channelsSampleRateHz)) {
             ////assert(samples_buffer_idx == channelsSampleRateHz && "should be at exactly the end, no further");
             if (UNLIKELY(samples_buffer_idx > channelsSampleRateHz)) {
-                ERRLOG("OOPS, possible overflow in speaker samples buffer ... samples_buffer_idx:%lu channelsSampleRateHz:%lu", samples_buffer_idx, channelsSampleRateHz);
+                ERRLOG("OOPS, possible overflow in speaker samples buffer ... samples_buffer_idx:%lu channelsSampleRateHz:%lu", (unsigned long)samples_buffer_idx, channelsSampleRateHz);
             }
         }
     }
